Added unit tests for CustomSqlMigrationSource. Automatic DB schema updates on registry startup now use SQL migrations.
