+++ test/unit/test-migrations/01_test_migr2.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS migr1
+(
+    id         int unsigned NOT NULL AUTO_INCREMENT,
+    PRIMARY KEY (id)
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4 COLLATE = utf8mb4_general_ci;
