+++ test/unit/test-migrations/02_test_migr3.js
@@ -1,0 +1,1 @@
+console.log('should not be found');
